pyproject.toml
[tool.poetry]
name = "mileage_service_drf"
version = "0.1.0"
description = ""
authors = ["KooHyunJung <ghj3160@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
black = "22.3.0"
djangorestframework = "^3.13.1"
django-stubs = "1.9.0"
mypy = "0.931"
isort = "5.10.1"
PyMySQL = "^1.0.2"
types-PyMySQL = "^1.0.19"
cryptography = "^37.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.isort]
profile = "black"